\name{Frames-internal}
        \alias{Lag1}
        \alias{Lag2}
        \alias{PELConfInt}
        \title{Internal Frame Functions}
        \description{
          Internal Frame functions
        }
        \details{
          These are not to be called by the user.
        }
        \keyword{internal} 