% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{PiklB}
\alias{PiklB}
\title{Matrix of inclusion probabilities for units selected in sample from frame B}
\usage{
PiklB
}
\description{
This dataset consists of a square matrix of dimension 135 with the first and second order inclusion probabilities
 for the units included in \eqn{s_B}.
}
\examples{
data(PiklB)
#Let choose the submatrix of inclusion probabilities for the first 5 units in sB.
PiklB[1:5, 1:5]
#Now, let select the first order inclusion probabilities
diag(PiklB)
}
\keyword{datasets}

