% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WeightsCalDF}
\alias{WeightsCalDF}
\title{g-weights for the dual frame calibration estimator}
\usage{
WeightsCalDF(ysA, ysB, pi_A, pi_B, domains_A, domains_B, N_A = NULL, N_B = NULL,
N_ab = NULL, xsAFrameA = NULL, xsBFrameA = NULL, xsAFrameB = NULL, xsBFrameB = NULL,
xsT = NULL, XA = NULL, XB = NULL, X = NULL, met = "linear")
}
\arguments{
\item{ysA}{A numeric vector of length \eqn{n_A} or a numeric matrix or data frame of dimensions \eqn{n_A} x \eqn{c} containing information about variable(s) of interest from \eqn{s_A}.}

\item{ysB}{A numeric vector of length \eqn{n_B} or a numeric matrix or data frame of dimensions \eqn{n_B} x \eqn{c} containing information about variable(s) of interest from \eqn{s_B}.}

\item{pi_A}{A numeric vector of length \eqn{n_A} or a square numeric matrix of dimension \eqn{n_A} containing first order or first and second order inclusion probabilities for units included in \eqn{s_A}.}

\item{pi_B}{A numeric vector of length \eqn{n_B} or a square numeric matrix of dimension \eqn{n_B} containing first order or first and second order inclusion probabilities for units included in \eqn{s_B}.}

\item{domains_A}{A character vector of length \eqn{n_A} indicating the domain each unit from \eqn{s_A} belongs to. Possible values are "a" and "ab".}

\item{domains_B}{A character vector of length \eqn{n_B} indicating the domain each unit from \eqn{s_B} belongs to. Possible values are "b" and "ba".}

\item{N_A}{(Optional) A numeric value indicating the size of frame A.}

\item{N_B}{(Optional) A numeric value indicating the size of frame B.}

\item{N_ab}{(Optional) A numeric value indicating the size of the overlap domain.}

\item{xsAFrameA}{(Optional) A numeric vector of length \eqn{n_A} or a numeric matrix or data frame of dimensions \eqn{n_A} x \eqn{m_A}, with \eqn{m_A} the number of auxiliary variables in frame A, containing auxiliary information in frame A for units included in \eqn{s_A}.}

\item{xsBFrameA}{(Optional) A numeric vector of length \eqn{n_B} or a numeric matrix or data frame of dimensions \eqn{n_B} x \eqn{m_A}, with \eqn{m_A} the number of auxiliary variables in frame A, containing auxiliary information in frame A for units included in \eqn{s_B}. For units in domain \eqn{b}, these values are 0.}

\item{xsAFrameB}{(Optional) A numeric vector of length \eqn{n_A} or a numeric matrix or data frame of dimensions \eqn{n_A} x \eqn{m_B}, with \eqn{m_B} the number of auxiliary variables in frame B, containing auxiliary information in frame B for units included in \eqn{s_A}. For units in domain \eqn{a}, these values are 0.}

\item{xsBFrameB}{(Optional) A numeric vector of length \eqn{n_B} or a numeric matrix or data frame of dimensions \eqn{n_B} x \eqn{m_B}, with \eqn{m_B} the number of auxiliary variables in frame B, containing auxiliary information in frame B for units included in \eqn{s_B}.}

\item{xsT}{(Optional) A numeric vector of length \eqn{n} or a numeric matrix or data frame of dimensions \eqn{n} x \eqn{m_T}, with \eqn{m_T} the number of auxiliary variables in both frames, containing auxiliary information for all units in the entire sample \eqn{s = s_A \cup s_B}.}

\item{XA}{(Optional) A numeric value or vector of length \eqn{m_A}, with \eqn{m_A} the number of auxiliary variables in frame A, indicating the population totals for the auxiliary variables considered in frame A.}

\item{XB}{(Optional) A numeric value or vector of length \eqn{m_B}, with \eqn{m_B} the number of auxiliary variables in frame B, indicating the population totals for the auxiliary variables considered in frame B.}

\item{X}{(Optional) A numeric value or vector of length \eqn{m_T}, with \eqn{m_T} the number of auxiliary variables in both frames, indicating the population totals for the auxiliary variables considered in both frames.}

\item{met}{(Optional) A character vector indicating the distance that must be used in calibration process. Possible values are "linear", "raking" and "logit". Default is "linear".}
}
\value{
A numeric vector containing the g-weights for the dual frame calibration estimator.
}
\description{
Computes the g-weights for the dual frame calibration estimator.
}
\details{
Function provides g-weights in following scenarios:
 \itemize{
	\item There is not any additional auxiliary variable
     \itemize{
     	\item \eqn{N_A, N_B} and \eqn{N_{ab}} unknown
		\item \eqn{N_{ab}} known and \eqn{N_A} and \eqn{N_B} unknown
             \item \eqn{N_A} and \eqn{N_B} known and \eqn{N_{ab}} unknown
             \item \eqn{N_A, N_B} and \eqn{N_{ab}} known
     }
     \item At least, one additional auxiliary variable is available
     \itemize{
		\item \eqn{N_{ab}} known and \eqn{N_A} and \eqn{N_B} unknown
             \item \eqn{N_A} and \eqn{N_B} known and \eqn{N_{ab}} unknown
             \item \eqn{N_A, N_B} and \eqn{N_{ab}} known
     }
 }
}
\examples{
data(HouseholdsA)
dataA <- attach(HouseholdsA)
detach(HouseholdsA)
data(HouseholdsB)
dataB <- attach(HouseholdsB)
detach(HouseholdsB)
data(PiklA)
data(PiklB)

#Let calculate g-weights for the dual frame calibration estimator for variable Feeding,
#without considering any auxiliary information
WeightsCalDF(dataA$Feeding, dataB$Feeding, PiklA, PiklB, dataA$Domain, dataB$Domain)

#Now, let calculate g-weights for the dual frame calibration estimator for variable Clothing
#when the frame sizes and the overlap domain size are known
WeightsCalDF(dataA$Clothing, dataB$Clothing, PiklA, PiklB, dataA$Domain, dataB$Domain,
N_A = 1735, N_B = 1191, N_ab = 601)

#Finally, let calculate g-weights for the dual frame calibration estimator
#for variable Feeding, considering Income as auxiliary variable in frame A
#and Metres2 as auxiliary variable in frame B and with frame sizes and overlap
#domain size known.
WeightsCalDF(dataA$Feeding, dataB$Feeding, PiklA, PiklB, dataA$Domain, dataB$Domain,
N_A = 1735, N_B =  1191, N_ab = 601, xsAFrameA = dataA$Income, xsBFrameA = dataB$Income,
xsAFrameB = dataA$Metres2, xsBFrameB = dataB$Metres2, XA = 4300260, XB = 176553)
}
\references{
Ranalli, M. G., Arcos, A., Rueda, M. and Teodoro, A. (2013)
 \emph{Calibration estimationn in dual frame surveys}. arXiv:1312.0761 [stat.ME]

Deville, J. C., S\"arndal, C. E. (1992)
 \emph{Calibration estimators in survey sampling.}
 Journal of the American Statistical Association, 87, 376 - 382
}

